requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
discord==2.3.2
discord.py==2.4.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
frozenlist==1.4.1
idna==3.8
jieba3k==0.35.1
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.2.2
multidict==6.0.5
newspaper3k==0.2.8
nltk==3.9.1
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
urllib3==2.2.2
yarl==1.9.7
